style: extract SidebarMainMissionControl
